Move RestrictedManager to sql_db
